Move pch.h into Common/ChairLoader
