+++ ChairLoader/ChairloaderGui.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include "ChairloaderGUI.h"
 
 ChairloaderGui *gui = nullptr;
