+++ ChairLoader/EntityManager.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include "EntityManager.h"
 #include <Prey/GameDll/ark/player/ArkPlayer.h>
