+++ ChairLoader/HeaderTestFile.cpp
@@ -3,5 +3,3 @@
 // You can compile just one file via Build -> Compile in Visual Studio.
 // This removes interference from other .cpp or .h files and keeps error list clean.
 //
-
-#include "pch.h"
