+++ ChairLoader/OverlayLog.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include "OverlayLog.h"
 #include "ChairLoader.h"
