+++ ChairLoader/PlayerManager.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include "PlayerManager.h"
 
 #include "Logging.h"
