+++ ChairLoader/dllmain.cpp
@@ -1,5 +1,4 @@
 // dllmain.cpp : Defines the entry point for the DLL application.
-#include "pch.h"
 #include <string>
 #include "mem.h"
 #include "ChairloaderGui.h"
