+++ Common/ChairLoader/PreyFunction.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include <ChairLoader/PreyFunction.h>
 #include <Prey/CryCore/Platform/CryWindows.h>
 #include <detours/detours.h>
