+++ Common/Logging.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include <Prey/CrySystem/ILog.h>
 #include <Logging.h>
