+++ Common/Prey/ArkAbilityLibrary.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "pch.h"
 #include <string>
 #include <map>
 class ArkAbilityLibrary {
