+++ Common/VTableHook.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include <cstdlib>
 #include <cstring>
 #include "VTableHook.h"
