+++ Common/imgui/imgui_draw.cpp
@@ -21,7 +21,6 @@ Index of this file:
 // [SECTION] Default font data (ProggyClean.ttf)
 
 */
-#include "pch.h"
 #if defined(_MSC_VER) && !defined(_CRT_SECURE_NO_WARNINGS)
 #define _CRT_SECURE_NO_WARNINGS
 #endif
