+++ Common/imgui/misc/cpp/imgui_stdlib.cpp
@@ -3,8 +3,7 @@
 
 // Changelog:
 // - v0.10: Initial version. Added InputText() / InputTextMultiline() calls with std::string
-#include "pch.h"
-#include "imgui.h"
+#include <imgui/imgui.h>
 #include "imgui_stdlib.h"
 
 struct InputTextCallback_UserData
