+++ Common/imgui/misc/fonts/binary_to_compressed_c.cpp
@@ -21,7 +21,6 @@
 //   # binary_to_compressed_c.exe -base85 myfont.ttf MyFont > myfont.cpp
 
 #define _CRT_SECURE_NO_WARNINGS
-#include "pch.h"
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
