+++ Common/mem.cpp
@@ -1,4 +1,3 @@
-#include "pch.h"
 #include "mem.h"
 #include <Prey/CryCore/Platform/CryWindows.h>
